+++ extensions/msal-node-extensions/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @azure/msal-node-extensions
 
-This log was last generated on Fri, 11 Aug 2023 19:00:44 GMT and should not be manually modified.
+This log was last generated on Fri, 18 Aug 2023 18:40:02 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.0.2
+
+Fri, 18 Aug 2023 18:40:02 GMT
+
+### Patches
+
+- Bump @azure/msal-common to v14.0.2
 
 ## 1.0.1